+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/key/KeyboardShortcuts.java
@@ -11,32 +11,34 @@ import java.util.function.Function;
 import static javax.swing.KeyStroke.getKeyStroke;
 
 /**
- * Holds keyboard shortcut keyStrokes.
- * @param <T> the shortcut key
+ * Holds keyboard shortcut keyStrokes, mapped to enum based shortcut keys.
+ * @param <T> the shortcut key type
  * @see #keyboardShortcuts(Class, Function)
  */
 public interface KeyboardShortcuts<T extends Enum<T>> {
 
   /**
    * @param keyboardShortcut the shortcut key
-   * @return the Value controlling the key stroke for the given shortcut key
+   * @return the {@link Value} controlling the key stroke for the given shortcut key
    */
   Value<KeyStroke> keyStroke(T keyboardShortcut);
 
   /**
    * @param shortcutKeyClass the shortcut key class
-   * @param defaults provides the default keystroke for each shortcut key
+   * @param defaultKeyStrokes provides the default keystroke for each shortcut key
    * @return a new {@link KeyboardShortcuts} instance
    * @param <T> the shortcut key type
+   * @throws IllegalStateException in case the default keyStrokes function does not provide keyStrokes for all shortcut keys
    */
-  static <T extends Enum<T>> KeyboardShortcuts<T> keyboardShortcuts(Class<T> shortcutKeyClass, Function<T, KeyStroke> defaults) {
-    return new DefaultKeyboardShortcuts<>(shortcutKeyClass, defaults);
+  static <T extends Enum<T>> KeyboardShortcuts<T> keyboardShortcuts(Class<T> shortcutKeyClass, Function<T, KeyStroke> defaultKeyStrokes) {
+    return new DefaultKeyboardShortcuts<>(shortcutKeyClass, defaultKeyStrokes);
   }
 
   /**
    * Creates a {@link KeyStroke} with the given keyCode and no modifiers.
    * @param keyCode the key code
    * @return a keystroke value
+   * @see KeyStroke#getKeyStroke(int, int)
    */
   static KeyStroke keyStroke(int keyCode) {
     return keyStroke(keyCode, 0);
@@ -47,6 +49,7 @@ public interface KeyboardShortcuts<T extends Enum<T>> {
    * @param keyCode the key code
    * @param modifiers the modifiers
    * @return a keystroke value
+   * @see KeyStroke#getKeyStroke(int, int)
    */
   static KeyStroke keyStroke(int keyCode, int modifiers) {
     return getKeyStroke(keyCode, modifiers);