+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - KeyboardShortcut added, first attempt at configurable keyboard shortcuts.
 - KeyboardShortcut replaced with KeyboardShortcuts, second iteration of configurable keyboard shortcuts.
 - KeyEvents and TransferFocusOnEnter moved to key package.
+- KeyboardShortcuts factory method now throws exception in case of a missing default shortcut keyStroke.
 ### is.codion.framework.domain
 - IdentityKeyGenerator now throws exception in case the generated keys result set is empty.
 ### is.codion.framework.model